requirements.txt
azure-ai-documentintelligence==1.0.0b4
azure-core>=1.30.0
azure-identity>=1.16.0
python-dotenv>=1.0.0
aiohttp>=3.8.0